extract initStartingAreaBox
